Make integration tests shareable across VMs
