+++ examples/morpheusvm/tests/workload/workload.go
@@ -1,0 +1,150 @@
+// Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package workload
+
+import (
+	"context"
+	"fmt"
+	"math"
+	"time"
+
+	"github.com/ava-labs/avalanchego/ids"
+
+	"github.com/ava-labs/hypersdk/auth"
+	"github.com/ava-labs/hypersdk/chain"
+	"github.com/ava-labs/hypersdk/codec"
+	"github.com/ava-labs/hypersdk/crypto/ed25519"
+	"github.com/ava-labs/hypersdk/examples/morpheusvm/actions"
+	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
+	"github.com/ava-labs/hypersdk/examples/morpheusvm/genesis"
+	"github.com/ava-labs/hypersdk/fees"
+	"github.com/ava-labs/hypersdk/rpc"
+	"github.com/ava-labs/hypersdk/tests/workload"
+
+	lrpc "github.com/ava-labs/hypersdk/examples/morpheusvm/rpc"
+)
+
+var (
+	_ workload.TxWorkloadFactory  = (*workloadFactory)(nil)
+	_ workload.TxWorkloadIterator = (*simpleTxWorkload)(nil)
+)
+
+type workloadFactory struct {
+	factory *auth.ED25519Factory
+}
+
+func New(minBlockGap int64) (*genesis.Genesis, workload.TxWorkloadFactory, error) {
+	// Load default pk
+	prefundedAddrStr := "morpheus1qrzvk4zlwj9zsacqgtufx7zvapd3quufqpxk5rsdd4633m4wz2fdjk97rwu"
+	privBytes, err := codec.LoadHex(
+		"323b1d8f4eed5f0da9da93071b034f2dce9d2d22692c172f3cb252a64ddfafd01b057de320297c29ad0c1f589ea216869cf1938d88c9fbd70d6748323dbf2fa7", //nolint:lll
+		ed25519.PrivateKeyLen,
+	)
+	if err != nil {
+		return nil, nil, err
+	}
+	gen := genesis.Default()
+	// Set WindowTargetUnits to MaxUint64 for all dimensions to iterate full mempool during block building.
+	gen.WindowTargetUnits = fees.Dimensions{math.MaxUint64, math.MaxUint64, math.MaxUint64, math.MaxUint64, math.MaxUint64}
+	// Set all lmiits to MaxUint64 to avoid limiting block size for all dimensions except bandwidth. Must limit bandwidth to avoid building
+	// a block that exceeds the maximum size allowed by AvalancheGo.
+	gen.MaxBlockUnits = fees.Dimensions{1800000, math.MaxUint64, math.MaxUint64, math.MaxUint64, math.MaxUint64}
+	gen.MinBlockGap = minBlockGap
+	gen.CustomAllocation = []*genesis.CustomAllocation{
+		{
+			Address: prefundedAddrStr,
+			Balance: 10_000_000_000_000,
+		},
+	}
+
+	priv := ed25519.PrivateKey(privBytes)
+	factory := auth.NewED25519Factory(priv)
+
+	return gen, &workloadFactory{
+		factory: factory,
+	}, nil
+}
+
+func (f *workloadFactory) NewWorkloads(uri string) ([]workload.TxWorkloadIterator, error) {
+	basicTxWorkload, err := f.NewSizedTxWorkload(uri, 1)
+	return []workload.TxWorkloadIterator{basicTxWorkload}, err
+}
+
+func (f *workloadFactory) NewSizedTxWorkload(uri string, size int) (workload.TxWorkloadIterator, error) {
+	cli := rpc.NewJSONRPCClient(uri)
+	networkID, _, blockchainID, err := cli.Network(context.Background())
+	if err != nil {
+		return nil, err
+	}
+	lcli := lrpc.NewJSONRPCClient(uri, networkID, blockchainID)
+	return &simpleTxWorkload{
+		factory: f.factory,
+		cli:     cli,
+		lcli:    lcli,
+		size:    size,
+	}, nil
+}
+
+type simpleTxWorkload struct {
+	factory   *auth.ED25519Factory
+	cli       *rpc.JSONRPCClient
+	lcli      *lrpc.JSONRPCClient
+	networkID uint32
+	chainID   ids.ID
+	count     int
+	size      int
+}
+
+func (g *simpleTxWorkload) Next() bool {
+	return g.count < g.size
+}
+
+func (g *simpleTxWorkload) GenerateTxWithAssertion(ctx context.Context) (*chain.Transaction, workload.TxAssertion, error) {
+	g.count++
+	other, err := ed25519.GeneratePrivateKey()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	aother := auth.NewED25519Address(other.PublicKey())
+	aotherStr := codec.MustAddressBech32(consts.HRP, aother)
+	parser, err := g.lcli.Parser(ctx)
+	if err != nil {
+		return nil, nil, err
+	}
+	_, tx, _, err := g.cli.GenerateTransaction(
+		ctx,
+		parser,
+		[]chain.Action{&actions.Transfer{
+			To:    aother,
+			Value: 1,
+		}},
+		g.factory,
+	)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return tx, func(ctx context.Context, uri string) error {
+		ctx, cancel := context.WithTimeout(ctx, time.Second)
+		defer cancel()
+
+		lcli := lrpc.NewJSONRPCClient(uri, g.networkID, g.chainID)
+		success, _, err := lcli.WaitForTransaction(ctx, tx.ID())
+		if err != nil {
+			return fmt.Errorf("failed to wait for tx %s: %w", tx.ID(), err)
+		}
+		if !success {
+			return fmt.Errorf("tx %s not accepted", tx.ID())
+		}
+		balance, err := lcli.Balance(ctx, aotherStr)
+		if err != nil {
+			return fmt.Errorf("failed to get balance of %s: %w", aotherStr, err)
+		}
+		if balance != 1 {
+			return fmt.Errorf("expected balance of 1, got %d", balance)
+		}
+		return nil
+	}, nil
+}
