+++ tests/workload/transactions.go
@@ -23,6 +23,8 @@ type TxWorkloadFactory interface {
 	NewSizedTxWorkload(uri string, size int) (TxWorkloadIterator, error)
 }
 
+type TxAssertion func(context.Context, string) error
+
 // TxWorkloadIterator provides an interface for generating a sequence of transactions and corresponding assertions.
 // The caller must proceed in the following sequence:
 // 1. Next
@@ -44,7 +46,7 @@ type TxWorkloadIterator interface {
 	// GenerateTxWithAssertion generates a new transaction and an assertion function that confirms
 	// 1. The tx was accepted on the provided URI
 	// 2. The state was updated as expected according to the provided URI
-	GenerateTxWithAssertion(context.Context) (*chain.Transaction, func(ctx context.Context, uri string) error, error)
+	GenerateTxWithAssertion(context.Context) (*chain.Transaction, TxAssertion, error)
 }
 
 func ExecuteWorkload(ctx context.Context, require *require.Assertions, uris []string, generator TxWorkloadIterator) {
